requirements.txt
braintrust>=0.5.6
autoevals>=0.1.0
elevenlabs>=2.36.1
python-dotenv>=1.0.0
requests>=2.31.0
openai-whisper>=20240930
